refactor: move helpers to helpers.ts
