+++ packages/ts-endpoint/src/Endpoint/__typetests__/withHeaders.typespec.ts
@@ -10,7 +10,6 @@ const endpoint = Endpoint({
   Method: 'GET',
   getPath: ({ id }) => `users/${id}/crayons`,
   Output: t.type({ crayons: t.array(t.string) }),
-  Opts: { stringifyBody: true },
 });
 
 const noHeaderEndpoint = Endpoint({
@@ -20,7 +19,6 @@ const noHeaderEndpoint = Endpoint({
   Method: 'GET',
   getPath: ({ id }) => `users/${id}/crayons`,
   Output: t.type({ crayons: t.array(t.string) }),
-  Opts: { stringifyBody: true },
 });
 
 const authenticatedEndpoint = withAuth(endpoint);
@@ -38,10 +36,9 @@ authenticatedNoHeaderEndpoint.Input.Headers.props.authorization;
 // @dts-jest:fail:snap when endpoint is not wrapped you must not pass the header
 noHeaderEndpoint.Input.Headers?.authorization;
 
-
 const withHeadersEndpoint = withHeaders({
-  foo: t.string
-})(endpoint)
+  foo: t.string,
+})(endpoint);
 
 // @dts-jest:pass:snap when endpoint is wrapped other headers are still present
 withHeadersEndpoint.Input.Headers.props.id;
@@ -49,8 +46,8 @@ withHeadersEndpoint.Input.Headers.props.id;
 withHeadersEndpoint.Input.Headers.props.foo;
 
 const withHeadersNoHeaderEndpoint = withHeaders({
-  foo: t.string
-})(noHeaderEndpoint)
+  foo: t.string,
+})(noHeaderEndpoint);
 
 // @dts-jest:pass:snap when endpoint with no previous headers is wrapped, foo header is present
 withHeadersNoHeaderEndpoint.Input.Headers.props.foo;
