+++ packages/ts-endpoint/src/Endpoint/helpers.ts
@@ -1,10 +1,12 @@
 import * as t from 'io-ts';
 import { EndpointInstance } from '.';
+import { RequiredKeys } from 'typelevel-ts';
+
+export const addSlash = (s: string) => (s.substr(0, 1) === '/' ? s : `/${s}`);
 
 export type TypeOfEndpointInstance<E extends EndpointInstance<any>> = {
   getPath: E['getPath'];
   getStaticPath: E['getStaticPath'];
-  Opts: E['Opts'];
   Method: E['Method'];
   Output: t.TypeOf<E['Output']>;
   Errors: {
@@ -18,3 +20,13 @@ export type TypeOfEndpointInstance<E extends EndpointInstance<any>> = {
       : never;
   };
 };
+
+export type DecodedPropsType<P> = P extends {} ? { [k in RequiredKeys<P>]: t.TypeOf<P[k]> } : never;
+export type EncodedPropsType<P> = P extends {}
+  ? { [k in RequiredKeys<P>]: t.OutputOf<P[k]> }
+  : never;
+
+export type DecodedInput<E extends EndpointInstance<any>> = DecodedPropsType<E['Input']>;
+export type EncodedInput<E extends EndpointInstance<any>> = EncodedPropsType<E['Input']>;
+export type DecodedOutput<E extends EndpointInstance<any>> = t.TypeOf<E['Output']>;
+export type EncodedOutput<E extends EndpointInstance<any>> = t.TypeOf<E['Output']>;
